fixup: [BEAM-5878] renamed test to have appropriate suffix not to run in Python 2 and avoid using exec
